+++ Engine/Source/Developer/Profiler/Private/Widgets/SDataGraph.cpp
@@ -270,7 +270,7 @@ void SDataGraph::UpdateState()
 			NumDataPoints = (int32)FirstTrackedStat->GraphDataSource->GetNumFrames();
 		}
 
-		NumVisiblePoints = FMath::Max( 0, FMath::TruncToInt(ThisGeometry.Size.X) / DistanceBetweenPoints );
+		NumVisiblePoints = FMath::Max( 0, static_cast<int32>(FMath::TruncToInt(ThisGeometry.Size.X)) / DistanceBetweenPoints );
 		// GraphOffset - Updated by OnMouseMove or by ScrollTo
 		GraphOffset = FMath::Clamp( GraphOffset, 0, FMath::Max(NumDataPoints-NumVisiblePoints,0) );
 		
@@ -306,7 +306,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 
 	/** Width of the alloted geometry that is used to draw a data graph. */
 	const float AreaX0 = 0.0f;
-	const float AreaX1 = AllottedGeometry.GetLocalSize().X;
+	const float AreaX1 = static_cast<float>(AllottedGeometry.GetLocalSize().X);
 
 	// Draw background.
 	FSlateDrawElement::MakeBox
@@ -337,7 +337,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 		SCOPE_CYCLE_COUNTER(STAT_DG_OnPaint);
 
 		const FTrackedStat& TrackedStat = *It.Value();
-		const float GraphYScale = AllottedGeometry.GetLocalSize().Y/ScaleY;
+		const float GraphYScale = static_cast<float>(AllottedGeometry.GetLocalSize().Y) / ScaleY;
 		
 		const float UnitTypeScale = TrackedStat.GraphDataSource->GetSampleType() != EProfilerSampleTypes::Memory ? 1.0f : CounterToTimeScale;
 		const float TimeAccuracyMS = FTimeAccuracy::AsFrameTime( TimeBasedAccuracy );
@@ -427,8 +427,8 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 				for (float GraphStartTimeMS = GraphOffsetMS; GraphStartTimeMS < GraphRangeEndMS; GraphStartTimeMS += TimeAccuracyMS)
 				{
 					const float Value = GraphDataSource->GetValueFromTimeRange( GraphStartTimeMS, GraphStartTimeMS + TimeAccuracyMS );
-					const float XPos = DistanceBetweenPoints*GraphPoints.Num();
-					const float YPos = FMath::Clamp( AllottedGeometry.GetLocalSize().Y - GraphYScale*Value*UnitTypeScale, 0.0f, AllottedGeometry.GetLocalSize().Y );
+					const float XPos = static_cast<float>(DistanceBetweenPoints * GraphPoints.Num());
+					const float YPos = FMath::Clamp( static_cast<float>(AllottedGeometry.GetLocalSize().Y) - GraphYScale*Value*UnitTypeScale, 0.0f, static_cast<float>(AllottedGeometry.GetLocalSize().Y) );
 					new (GraphPoints)FVector2D( XPos, YPos );
 				}
 
@@ -458,7 +458,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 				{
 					const float Value = GraphDataSource->GetValueFromIndex( GraphStartIndex );
 					const float XPos = DistanceBetweenPoints*(float)GraphPoints.Num();
-					const float YPos = FMath::Clamp( AllottedGeometry.Size.Y - GraphYScale*Value*UnitTypeScale, 0.0f, AllottedGeometry.Size.Y );
+					const float YPos = FMath::Clamp( static_cast<float>(AllottedGeometry.Size.Y) - GraphYScale*Value*UnitTypeScale, 0.0f, static_cast<float>(AllottedGeometry.Size.Y) );
 					GraphPoints.Add( FVector2D(XPos,YPos) );
 				}
  
@@ -481,7 +481,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 
 
 	FSlateFontInfo SummaryFont = FCoreStyle::GetDefaultFontStyle("Regular", 8);
-	const float MaxFontCharHeight = FontMeasureService->Measure( TEXT("!"), SummaryFont ).Y;
+	const float MaxFontCharHeight = static_cast<float>(FontMeasureService->Measure( TEXT("!"), SummaryFont ).Y);
 
 	// Draw graph annotations.
 	//const IDataProviderPtr DataProvider = StatIDToGraphDescriptionMapping.Num() > 0 ? StatIDToGraphDescriptionMapping.CreateConstIterator().Key().DataSource->GetDataProvider() : NULL;
@@ -514,7 +514,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 
 			for( int32 FrameIndex = FrameStartIndex; FrameIndex < FrameEndIndex; FrameIndex += AvgFrameRate )
 			{
-				const float MarkerPosX = (FrameIndex - GraphOffset) * DistanceBetweenPoints;
+				const float MarkerPosX = static_cast<float>((FrameIndex - GraphOffset) * DistanceBetweenPoints);
 				const float ElapsedFrameTimeMS = DataProvider->GetElapsedFrameTimeMS( FrameIndex );
 
 				LinePoints.Add( FVector2D(MarkerPosX, 0.0) );
@@ -574,7 +574,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 
 			for( int32 FrameIndex = FrameStartIndex; FrameIndex < FrameEndIndex; FrameIndex += AvgFrameRate )
 			{
-				const float MarkerPosX = (FrameIndex - GraphOffset) * DistanceBetweenPoints;
+				const float MarkerPosX = static_cast<float>((FrameIndex - GraphOffset) * DistanceBetweenPoints);
 				const float ElapsedFrameTimeMS = FrameIndex * FTimeAccuracy::AsFrameTime( TimeBasedAccuracy );
 				const int32 ElapsedFrameTime = FMath::Max( FMath::RoundToInt( ElapsedFrameTimeMS * 0.001f ) - 1, 0 );
 				const int32 AccumulatedFrameCounter = bCanBeDisplayedAsMulti ? FrameIndex : DataProvider->GetAccumulatedFrameCounter(ElapsedFrameTime);
@@ -631,7 +631,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 
 	//-----------------------------------------------------------------------------
 
-	const int32 MaxGridPixelSpacing = 160.0f;
+	const int32 MaxGridPixelSpacing = 160;
 
 	// Draw a horizontal lines every 150 pixels and draw a few basic lines
 	static const TArray<float> DefaultTimeValueHints = TArrayBuilder<float>()
@@ -658,7 +658,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 	const float MinTimeValue = 0.0f;
 	const float MaxTimeValue = ScaleY;
 	const float TimeValueGraphScale = MaxTimeValue / SecondaryIndicators;
-	const float TimeValueToGraph = AllottedGeometry.GetLocalSize().Y/MaxTimeValue;
+	const float TimeValueToGraph = static_cast<float>(AllottedGeometry.GetLocalSize().Y) / MaxTimeValue;
 
 	for( int32 SecondaryIndex = 1; SecondaryIndex <= SecondaryIndicators; SecondaryIndex++ )
 	{
@@ -755,7 +755,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 			continue;
 		}
 
-		const float MarkerPosY = AllottedGeometry.GetLocalSize().Y - TimeValue*TimeValueToGraph;
+		const float MarkerPosY = static_cast<float>(AllottedGeometry.GetLocalSize().Y) - TimeValue * TimeValueToGraph;
 
 		// Check if this hint should be drawn as the basic hint.
 		const FLinearColor* BasicHintColor = DefaultTimeValueHintColors.Find( TimeValue );
@@ -807,7 +807,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 
 		// Right	-	Values in human readable string, for the non-hierarchical samples
 		const FString CounterValueStr = FString::Printf( TEXT("%.1f KB"), TimeValue/CounterToTimeScale );
-		const float RightValueSizeX = FontMeasureService->Measure( CounterValueStr, SummaryFont ).X;
+		const float RightValueSizeX = static_cast<float>(FontMeasureService->Measure( CounterValueStr, SummaryFont ).X);
 		FSlateDrawElement::MakeText
 		(
 			OutDrawElements, 
@@ -823,7 +823,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 	// Draw selected frames markers.
 	{
 		LayerId++;
-		const float LocalGraphOffset = GraphOffset * DistanceBetweenPoints;
+		const float LocalGraphOffset = static_cast<float>(GraphOffset * DistanceBetweenPoints);
 		const float LocalGraphSelectionX0 = FrameIndices[0]*DistanceBetweenPoints - LocalGraphOffset;
 		const float LocalGraphSelectionX1 = FrameIndices[1]*DistanceBetweenPoints - LocalGraphOffset;
 		const float LocalGraphSelectionX[2] = { LocalGraphSelectionX0, LocalGraphSelectionX1 };
@@ -874,7 +874,7 @@ int32 SDataGraph::OnPaint( const FPaintArgs& Args, const FGeometry& AllottedGeom
 		LayerId++;
 
 		const int32 LocalPosition = HoveredFrameIndex - GraphOffset;
-		const float LocalPositionGraphX = LocalPosition * DistanceBetweenPoints;
+		const float LocalPositionGraphX = static_cast<float>(LocalPosition * DistanceBetweenPoints);
 
 		FSlateDrawElement::MakeBox
 		(
@@ -1003,7 +1003,7 @@ FReply SDataGraph::OnMouseButtonDown( const FGeometry& MyGeometry, const FPointe
 		if( NumDataPoints > 0 )
 		{
 			// Capture mouse, so we can scroll outside this widget.
-			RealGraphOffset = GraphOffset;
+			RealGraphOffset = static_cast<float>(GraphOffset);
 			Reply = FReply::Handled().CaptureMouse(SharedThis(this));
 		}
 	}
@@ -1095,7 +1095,7 @@ FReply SDataGraph::OnMouseMove( const FGeometry& MyGeometry, const FPointerEvent
 		if( HasMouseCapture() && !MouseEvent.GetCursorDelta().IsZero() )
 		{
 			bIsRMB_Scrolling = true;
-			const float ScrollByAmount = -MouseEvent.GetCursorDelta().X * (1.0f/DistanceBetweenPoints);
+			const float ScrollByAmount = -static_cast<float>(MouseEvent.GetCursorDelta().X) * (1.0f/DistanceBetweenPoints);
 			RealGraphOffset += ScrollByAmount;
 
 			GraphOffset = FMath::Clamp( FMath::TruncToInt( RealGraphOffset ), 0, FMath::Max(NumDataPoints-NumVisiblePoints,0) );
@@ -1111,7 +1111,7 @@ FReply SDataGraph::OnMouseMove( const FGeometry& MyGeometry, const FPointerEvent
 const int32 SDataGraph::CalculateFrameIndex( const FVector2D InMousePosition ) const
 {
 	const float ScaleX = 1.0f/DistanceBetweenPoints;
-	const int32 MousePositionOffset = FMath::TruncToInt( (InMousePosition.X+HalfGraphMarkerWidth) * ScaleX );
+	const int32 MousePositionOffset = FMath::TruncToInt( (static_cast<float>(InMousePosition.X) + HalfGraphMarkerWidth) * ScaleX );
 	return FMath::Clamp( GraphOffset+MousePositionOffset, 0, NumDataPoints-1 );
 }
